moved function to build sts to separate package, fixed unit tests
