--- controllers/mongodb_authentication.go
@@ -1,22 +1,0 @@
-package controllers
-
-import (
-	"github.com/mongodb/mongodb-kubernetes-operator/pkg/kube/podtemplatespec"
-	"github.com/mongodb/mongodb-kubernetes-operator/pkg/kube/statefulset"
-)
-
-// buildScramPodSpecModification will add the keyfile volume to the podTemplateSpec
-// the keyfile is owned by the agent, and is required to have 0600 permissions.
-func buildScramPodSpecModification(mdb MongoDBStatefulSetOwner) podtemplatespec.Modification {
-	mode := int32(0600)
-	keyFileNsName := mdb.GetAgentKeyfileSecretNamespacedName()
-	keyFileVolume := statefulset.CreateVolumeFromSecret(keyFileNsName.Name, keyFileNsName.Name, statefulset.WithSecretDefaultMode(&mode))
-	keyFileVolumeVolumeMount := statefulset.CreateVolumeMount(keyFileVolume.Name, "/var/lib/mongodb-mms-automation/authentication", statefulset.WithReadOnly(false))
-	keyFileVolumeVolumeMountMongod := statefulset.CreateVolumeMount(keyFileVolume.Name, "/var/lib/mongodb-mms-automation/authentication", statefulset.WithReadOnly(false))
-
-	return podtemplatespec.Apply(
-		podtemplatespec.WithVolume(keyFileVolume),
-		podtemplatespec.WithVolumeMounts(agentName, keyFileVolumeVolumeMount),
-		podtemplatespec.WithVolumeMounts(mongodbName, keyFileVolumeVolumeMountMongod),
-	)
-}
